requirements.txt
streamlit
ultralytics
huggingface_hub
Pillow
numpy
torch
torchvision
transformers
requests
jinja2==3.1.4
